Big ol' commit with addition of support for GBA, Genesis, & N64 cartridge reading!
have some cleanup to do:
clean up sega dumping so don't need a page0/1
implement sega single reads

Add GBA to some of the common opcodes

dumping Don't think need addrH |= of mapper, but maybe this is key to
cleaning up first note..

gba, sega, n64 has extra NOPs, remove and test.

create pinport renames for sega pins, move mask defines to pinport.h

clean up comments for genesis page reads..

+++ firmware/source/gba.h
@@ -9,6 +9,6 @@ uint8_t gba_call( uint8_t opcode, uint8_t miscdata, uint16_t operand, uint8_t *r
 
 uint16_t gba_rd();
 void gba_latch_addr( uint16_t addr_lo, uint8_t addr_hi);
-uint8_t gba_page_rd( uint8_t *data, uint8_t len);
+uint8_t gba_page_rd( uint16_t *data, uint8_t len);
 
 #endif
